\name{input_2017}
\docType{data}
\alias{input_2017}
\title{Input Data for the 2017 ACIC Competition}
\description{
  Input data used in the 2017 Atlantic Causal Inference Competition, taken from the Collaborative Perinatal Project.
}
\usage{input_2017}
\format{A data frame consisting of 4302 observations and 58 covariates. Factor columns remain coded with hypothetically
        relevant substantive information such as race, while the numeric column labels have been discarded. Variables
        we chosen to correspond to possible confounders, instruments, and uncorrelated variables from a hypothetical
        twin study on the impact of birthweight on IQ.}
\source{
  Niswander, K. R. and Gordon, M. (1972) The Collaborative Perinatal Study of the National Institute of Neurological
  Diseases and Stroke: the women and their pregnancies. Philadelphia, PA: W.B. Saunders Company
  \url{https://www.archives.gov/research/electronic-records/nih.html}
}
\keyword{datasets}

